StepHitPoints: Remove leading '1' from the hitpoints edit display
The StepHitPoints component displays a leading '1' to the HP display
which works fine for level 1 characters but as the character's hit
dice increase this results in erroneous display e.g. 12d4 for a 2nd
level magic user.

This fix removes the leading '1'. This fixes the issue for levels 2+
but results in the HD for 1st level characters showing without a
number prefix.